+++ LotCoMClient/Models/Datasources/RecordParser.cs
@@ -1,3 +1,4 @@
+using System.Text.RegularExpressions;
 using LotCoMClient.Models.Exceptions;
 
 namespace LotCoMClient.Models.Datasources;
@@ -26,7 +27,16 @@ public static class RecordParser {
         if (SplitLine.Count < 7) {
             throw new RecordParseException();
         }
-        // prepare DataRecord properties
+        // peek the first element and test it as an IP Address using a Regex pattern
+        string? ScanAddress = null;
+        string AddressPattern = @"^\d\d?\d?\.\d\d?\d?\.\d\d?\d?\.\d\d?\d?$";
+        Regex AddressRegex = new Regex(AddressPattern);
+        if (AddressRegex.IsMatch(SplitLine[0])) {
+            // set the ScanAddress property using this field and pop it off of the front
+            ScanAddress = SplitLine[0];
+            SplitLine.RemoveAt(0);
+        }
+        // prepare remaining universal DataRecord properties
         Process RecordProcess;
         Part RecordPart;
         string Quantity = SplitLine[3];
@@ -34,6 +44,7 @@ public static class RecordParser {
         string LotNumber = "";
         string DeburrJBKNumber = "";
         string DieNumber = "";
+        string ModelNumber = "";
         string HeatNumber = "";
         string RecordDate = SplitLine[^3].Split("-")[0];
         string RecordTime = SplitLine[^3].Split("-")[1];
@@ -53,7 +64,6 @@ public static class RecordParser {
         }
         // parse the required variable data fields
         List<string> Requirements = RecordProcess.RequiredFields;
-        List<string> InnerFields = [JBKNumber, LotNumber, DeburrJBKNumber, DieNumber, HeatNumber];
         // references next parsable index after the static Quantity, position 4 (index 3); 
         // increments when a property is found to be required and is assigned a parsable index
         int _parsingIndex = 0;
@@ -85,6 +95,13 @@ public static class RecordParser {
             // increment to the next parsable index
             _parsingIndex += 1;
         }
+        // attempt to parse a Model number
+        if (Requirements.Contains("ModelNumber")) {
+            // assign the value of the current parsing index to the Model Number property
+            ModelNumber = SplitLine[4 + _parsingIndex];
+            // increment to the next parsable index
+            _parsingIndex += 1;
+        }
         // attempt to parse a Heat number
         if (Requirements.Contains("HeatNumber")) {
             // assign the value of the current parsing index to the Heat Number property
@@ -92,10 +109,9 @@ public static class RecordParser {
             // increment to the next parsable index
             _parsingIndex += 1;
         }
-
         // attempt to create a DataRecord from the parsed data
         try {
-            return new DataRecord(RecordProcess, RecordPart, Quantity, JBKNumber, LotNumber, DeburrJBKNumber, DieNumber, HeatNumber, RecordDate, RecordTime, RecordShift, OperatorID);
+            return new DataRecord(RecordProcess, RecordPart, Quantity, JBKNumber, LotNumber, DeburrJBKNumber, DieNumber, ModelNumber, HeatNumber, RecordDate, RecordTime, RecordShift, OperatorID, ScanAddress);
         // there was a problem constructing a DataRecord from the parsed data
         } catch {
             throw new RecordParseException();